% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\docType{data}
\name{admission.symptoms}
\alias{admission.symptoms}
\title{Heat admission.symptoms}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 23 rows and 2 columns.
}
\usage{
admission.symptoms
}
\description{
Create the Heat admission.symptoms
}
\keyword{datasets}
