% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{admission.to.icu.prep}
\alias{admission.to.icu.prep}
\title{Aggregate data for hospital admission to ICU admission}
\usage{
admission.to.icu.prep(input.tbl)
}
\arguments{
\item{input.tbl}{Input tibble (output of \code{data.preprocessing})}
}
\value{
A \code{tibble} containing the input data for the hospital admission to ICU plot
}
\description{
Aggregate data for hospital admission to ICU admission
}
