% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{age.pyramid.prep}
\alias{age.pyramid.prep}
\title{Aggregate data for age pyramid plot}
\usage{
age.pyramid.prep(input.tbl)
}
\arguments{
\item{input.tbl}{Input tibble (output of \code{process.all.data})}
}
\value{
A \code{tibble} containing the input data for the age pyramid plot
}
\description{
Aggregate data for age pyramid plot
}
