% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{comorbidity.prep}
\alias{comorbidity.prep}
\title{Prepare Table4. comorbidities}
\usage{
comorbidity.prep(input.tbl)
}
\arguments{
\item{input.tbl}{Input tibble (output of \code{data.preprocessing})}
}
\value{
A \code{tibble} containing the input data for the Patient characteristics table
}
\description{
Prepare Table4. comorbidities
}
