% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{comorbidity.prevalence.prep}
\alias{comorbidity.prevalence.prep}
\title{Aggregate data for comorbidity prevalence plot}
\usage{
comorbidity.prevalence.prep(input.tbl)
}
\arguments{
\item{input.tbl}{Input tibble (output of \code{data.preprocessing})}
}
\value{
A \code{tibble} containing the input data for the comorbidity prevalence plot
}
\description{
Aggregate data for comorbidity prevalence plot
}
