% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{comorbidity.upset.prep}
\alias{comorbidity.upset.prep}
\title{Aggregate data for comorbidities upset plot}
\usage{
comorbidity.upset.prep(input.tbl, max.comorbidities = 5)
}
\arguments{
\item{input.tbl}{Input tibble (output of \code{data.preprocessing})}

\item{max.comorbidities}{The plot will display only the n most common comorbidities, this parameter is n}
}
\value{
A \code{tibble} containing the input data for the comorbidities upset plot
}
\description{
Aggregate data for comorbidities upset plot
}
