% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PreprocessingFunction.R
\docType{data}
\name{date_pull}
\alias{date_pull}
\title{Preprocessing step for all aggregations. Currently: remaps outcome to death, discharge or NA, cuts age into 5-year age groups, and adds a year-epiweek column}
\format{
An object of class \code{Date} of length 1.
}
\usage{
date_pull
}
\arguments{
\item{input.tbl}{Input tibble (output of \code{process.all.data})}
}
\value{
A \code{tibble} intended for input into other aggregation functions (e.g. \code{age.pyramid.prep})
}
\description{
Preprocessing step for all aggregations. Currently: remaps outcome to death, discharge or NA, cuts age into 5-year age groups, and adds a year-epiweek column
}
\keyword{datasets}
