% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{func_plot_comorbid_asthma}
\alias{func_plot_comorbid_asthma}
\title{Comorbidities by age}
\usage{
func_plot_comorbid_asthma(input.tbl)
}
\arguments{
\item{input.tbl}{Input tibble (output of \code{data.preprocessing})}
}
\value{
A \code{tibble} containing the input data for the Comorbidities data
}
\description{
Comorbidities by age
}
