% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{func_plot_lab_crp}
\alias{func_plot_lab_crp}
\title{Box and whisker plots for laboratory results at hospital presentation stratified by age group.}
\usage{
func_plot_lab_crp(input.tbl)
}
\arguments{
\item{input.tbl}{Input tibble (output of \code{data.preprocessing})}
}
\value{
A \code{tibble} containing the input data for the lab data
}
\description{
Box and whisker plots for laboratory results at hospital presentation stratified by age group.
}
