% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{func_plot_symptoms_history_of_fever}
\alias{func_plot_symptoms_history_of_fever}
\title{symptoms by age}
\usage{
func_plot_symptoms_history_of_fever(input.tbl)
}
\arguments{
\item{input.tbl}{Input tibble (output of \code{data.preprocessing})}
}
\value{
A \code{tibble} containing the input data for the Comorbidities data
}
\description{
symptoms by age
}
