% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{func_plots_vs_resp}
\alias{func_plots_vs_resp}
\title{vital signs}
\usage{
func_plots_vs_resp(input.tbl)
}
\arguments{
\item{input.tbl}{Input tibble (output of \code{data.preprocessing})}
}
\value{
A \code{tibble} containing the input data for the vital sign tables plot
}
\description{
vital signs
}
