% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFunctions.R
\name{import.demographic.data}
\alias{import.demographic.data}
\title{Import demographic data}
\usage{
import.demographic.data(file.name, dtplyr.step)
}
\arguments{
\item{file.name}{Path of the demographics data file (CDISC format)}

\item{dtplyr.step}{Return the output as \code{dtplyr_step} to avoid unnecessary future calls to \code{as_tibble} or \code{as.data.table}}
}
\value{
Formatted demographic data as a tibble or \code{dtplyr_step}
}
\description{
Import demographic data
}
