% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFunctions.R, R/ImportFunctionsModular.R
\name{import.microb.data}
\alias{import.microb.data}
\title{Import microb data}
\usage{
import.microb.data(file.name, dtplyr.step = FALSE)

import.microb.data(file.name, dtplyr.step = FALSE)
}
\arguments{
\item{file.name}{Path of the microbio data file (CDISC format)}

\item{dtplyr.step}{Return the output as \code{dtplyr_step} to avoid unnecessary future calls to \code{as_tibble} or \code{as.data.table}}
}
\value{
Formatted demographic data as a tibble or \code{dtplyr_step}

Formatted demographic data as a tibble or \code{dtplyr_step}
}
\description{
Import microb data

Import microb data
}
