% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{key.times.prep}
\alias{key.times.prep}
\title{Prepare Table6. key times variable}
\usage{
key.times.prep(input.tbl)
}
\arguments{
\item{input.tbl}{Input tibble (output of \code{data.preprocessing})}
}
\value{
A \code{tibble} containing the input data for the Patient characteristics table
}
\description{
Prepare Table6. key times variable
}
