% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{length.of.stay.icu.prep}
\alias{length.of.stay.icu.prep}
\title{Aggregate data for length of stay within ICU}
\usage{
\method{length}{of.stay.icu.prep}(input.tbl)
}
\arguments{
\item{input.tbl}{Input tibble (output of \code{data.preprocessing})}
}
\value{
A \code{tibble} containing the input data for the length of stay within plot
}
\description{
Aggregate data for length of stay within ICU
}
