% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{length.of.stay.sex.prep}
\alias{length.of.stay.sex.prep}
\title{Aggregate data for hospital stay plot by sex}
\usage{
\method{length}{of.stay.sex.prep}(input.tbl)
}
\arguments{
\item{input.tbl}{Input tibble (output of \code{data.preprocessing})}
}
\value{
A \code{tibble} containing the input data for the hospital stay plot by sex
}
\description{
Aggregate data for hospital stay plot by sex
}
