% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{outcome.admission.date.prep}
\alias{outcome.admission.date.prep}
\title{Aggregate data for outcome by admission date plot}
\usage{
outcome.admission.date.prep(input.tbl)
}
\arguments{
\item{input.tbl}{Input tibble (output of \code{data.preprocessing})}
}
\value{
A \code{tibble} containing the input data for the outcome by admission date plot
}
\description{
Aggregate data for outcome by admission date plot
}
