% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{outcome.age.sex.prep}
\alias{outcome.age.sex.prep}
\title{Prepare Table2. Outcome by age and sex}
\usage{
outcome.age.sex.prep(input.tbl)
}
\arguments{
\item{input.tbl}{Input tibble (output of \code{data.preprocessing})}
}
\value{
A \code{tibble} containing the input data for the Patient characteristics table
}
\description{
Prepare Table2. Outcome by age and sex
}
