% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{patient.by.case.def.prep}
\alias{patient.by.case.def.prep}
\title{Case definitions data
From christiana's paper}
\usage{
patient.by.case.def.prep(input.tbl)
}
\description{
Case definitions data
From christiana's paper
}
