% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{patient.by.country.prep}
\alias{patient.by.country.prep}
\title{Aggregate data for patient by country}
\usage{
patient.by.country.prep(input.tbl)
}
\arguments{
\item{input.tbl}{Input tibble (output of \code{data.preprocessing})}
}
\value{
A \code{tibble} containing the input data for patient by country plot
}
\description{
Aggregate data for patient by country
}
