% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{patient.characteristic.prep}
\alias{patient.characteristic.prep}
\title{Prepare Table1. Patient characteristics}
\usage{
patient.characteristic.prep(input.tbl)
}
\arguments{
\item{input.tbl}{Input tibble (output of \code{data.preprocessing})}
}
\value{
A \code{tibble} containing the input data for the Patient characteristics table
}
\description{
Prepare Table1. Patient characteristics
}
