% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{patient.site.time.map.prep}
\alias{patient.site.time.map.prep}
\title{Aggregate data for case enrolment over time by site}
\usage{
patient.site.time.map.prep(input.tbl)
}
\arguments{
\item{input.tbl}{Input tibble (output of \code{data.preprocessing})}
}
\value{
A \code{tibble} containing the input data for the moving map
}
\description{
Aggregate data for case enrolment over time by site
}
