% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFunctions.R
\name{process.all.data}
\alias{process.all.data}
\title{Fully process data}
\usage{
process.all.data(
  demog.file.name,
  symptoms.file.name = NA,
  ICU.file.name = NA,
  treatment.file.name = NA,
  outcome.file.name = NA,
  minimum.treatments = 1000,
  dtplyr.step = FALSE
)
}
\arguments{
\item{demog.file.name}{Path of the demographics data file (CDISC format)}

\item{symptoms.file.name}{Path of the symptoms data file (CDISC format, optional)}

\item{ICU.file.name}{Path of the healthcare encounters data file (CDISC format, optional)}

\item{treatment.file.name}{Path of the intervention data file (CDISC format, optional)}

\item{outcome.file.name}{Path of the dispositions data file (CDISC format, optional)}

\item{minimum.treatments}{The minimum number of instances of a treatment required for inclusion as a column}

\item{dtplyr.step}{Return the output as \code{dtplyr_step} to avoid unnecessary future calls to \code{as_tibble} or \code{as.data.table}}
}
\value{
Formatted outcome data as a tibble or \code{dtplyr_step}
}
\description{
Fully process data
}
