% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFunctions.R
\name{process.common.treatment.data}
\alias{process.common.treatment.data}
\title{Process data on the most common treatments}
\usage{
process.common.treatment.data(input, minimum = 1000, dtplyr.step = FALSE)
}
\arguments{
\item{input}{Either the path of the interventions data file (CDISC format) or output of \code{process.treatment.data}}

\item{minimum}{The minimum number of times a treatment need appear to be considered "common"; default 1000.}

\item{dtplyr.step}{Return the output as \code{dtplyr_step} to avoid unnecessary future calls to \code{as_tibble} or \code{as.data.table}}
}
\value{
Formatted common treatment data (wide format) as a tibble or \code{dtplyr_step}
}
\description{
Process data on the most common treatments
}
