% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFunctions.R, R/ImportFunctionsModular.R
\name{process.comorbidity.data}
\alias{process.comorbidity.data}
\title{Process data on comorbidities}
\usage{
process.comorbidity.data(input, minimum = 100, dtplyr.step = FALSE)

process.comorbidity.data(input, minimum = 100, dtplyr.step = FALSE)
}
\arguments{
\item{input}{Either the path of the symptoms/comorbidities data file (CDISC format) or output of \code{import.symptom.and.comorbidity.data}}

\item{dtplyr.step}{Return the output as \code{dtplyr_step} to avoid unnecessary future calls to \code{as_tibble} or \code{as.data.table}}
}
\value{
Formatted comorbidity data as a tibble or \code{dtplyr_step}

Formatted comorbidity data as a tibble or \code{dtplyr_step}
}
\description{
Process data on comorbidities

Process data on comorbidities
}
