% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFunctions.R, R/ImportFunctionsModular.R
\name{process.ICU.data}
\alias{process.ICU.data}
\title{Process data on ICU admission}
\usage{
process.ICU.data(file.name, dtplyr.step = FALSE)

process.ICU.data(file.name, dtplyr.step = FALSE)
}
\arguments{
\item{file.name}{Path of the healthcare encounters data file (CDISC format)}

\item{dtplyr.step}{Return the output as \code{dtplyr_step} to avoid unnecessary future calls to \code{as_tibble} or \code{as.data.table}}
}
\value{
Formatted symptom data as a tibble or \code{dtplyr_step}

Formatted symptom data as a tibble or \code{dtplyr_step}
}
\description{
Process data on ICU admission

Process data on ICU admission
}
