% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFunctions.R
\name{process.IMV.NIV.ECMO.data}
\alias{process.IMV.NIV.ECMO.data}
\title{Process dates on IMV and NIV}
\usage{
process.IMV.NIV.ECMO.data(input, dtplyr.step = FALSE)
}
\arguments{
\item{input}{Either the path of the interventions data file (CDISC format) or output of \code{process.treatment.data}}

\item{dtplyr.step}{Return the output as \code{dtplyr_step} to avoid unnecessary future calls to \code{as_tibble} or \code{as.data.table}}
}
\value{
Formatted start (in) and end (out) dates for IMV and NIV treatment (wide format) as a tibble or \code{dtplyr_step}
}
\description{

}
\details{

}
