% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFunctions.R, R/ImportFunctionsModular.R
\name{process.pregnancy.data}
\alias{process.pregnancy.data}
\title{Process data on pregnancy (as comorbidity)}
\usage{
process.pregnancy.data(file.name, dtplyr.step = FALSE)

process.pregnancy.data(file.name, dtplyr.step = FALSE)
}
\arguments{
\item{file.name}{Path of the dispositions data file (CDISC format)}

\item{dtplyr.step}{Return the output as \code{dtplyr_step} to avoid unnecessary future calls to \code{as_tibble} or \code{as.data.table}}
}
\value{
Formatted pregnancy data as a tibble or \code{dtplyr_step}

Formatted pregnancy data as a tibble or \code{dtplyr_step}
}
\description{
Process data on pregnancy (as comorbidity)

Process data on pregnancy (as comorbidity)
}
