% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFunctions.R
\name{process.treatment.data}
\alias{process.treatment.data}
\title{Process data on treatments}
\usage{
process.treatment.data(file.name, dtplyr.step = FALSE)
}
\arguments{
\item{file.name}{Path of the intervention data file (CDISC format)}

\item{dtplyr.step}{Return the output as \code{dtplyr_step} to avoid unnecessary future calls to \code{as_tibble} or \code{as.data.table}}
}
\value{
Formatted treatment data (long format) as a tibble or \code{dtplyr_step}
}
\description{
Process data on treatments
}
