% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFunctionsModular.R
\name{process.treatment.icu.data}
\alias{process.treatment.icu.data}
\title{Process data on the most common icu treatments}
\usage{
process.treatment.icu.data(
  file.name,
  imp_icu,
  imp_dm,
  minimum = 10,
  dtplyr.step = FALSE
)
}
\arguments{
\item{minimum}{The minimum number of times a treatment need appear to be considered "common"; default 1000.}

\item{dtplyr.step}{Return the output as \code{dtplyr_step} to avoid unnecessary future calls to \code{as_tibble} or \code{as.data.table}}

\item{input}{Either the path of the interventions data file (CDISC format) or output of \code{process.treatment.data}}
}
\value{
Formatted common treatment data (wide format) as a tibble or \code{dtplyr_step}
}
\description{
Process data on the most common icu treatments
}
