% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFunctions.R, R/ImportFunctionsModular.R
\name{process.vital.sign.data}
\alias{process.vital.sign.data}
\title{Process data on vital sign}
\usage{
process.vital.sign.data(file.name, dtplyr.step = FALSE)

process.vital.sign.data(file.name, dtplyr.step = FALSE)
}
\arguments{
\item{file.name}{Path of the dispositions data file (CDISC format)}

\item{dtplyr.step}{Return the output as \code{dtplyr_step} to avoid unnecessary future calls to \code{as_tibble} or \code{as.data.table}}
}
\value{
Formatted vital sign (wide format) as a tibble or \code{dtplyr_step}

Formatted vital sign (wide format) as a tibble or \code{dtplyr_step}
}
\description{
Process data on vital sign

Process data on vital sign
}
