% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RandomizationFunction.R
\name{randomization}
\alias{randomization}
\title{Generate a randomized dataset based on a real wide dataset supplied which maintains key relationships.}
\usage{
randomization(import.tbl, var_identifier = "usubjid")
}
\arguments{
\item{var_identifier}{Vector of names of variables that are identifying variables (default = "usubjid")}

\item{data}{Dataframe to be used as a basis for a randomised dataset.}
}
\value{
data.frame with values synthesised based on the original dataset
}
\description{
Used to generate a randomized dataset that can be used to produce aggregated table with possible real data
}
