% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFunctions.R
\name{shared.data.import}
\alias{shared.data.import}
\title{Shared pre-processing of input CSV files}
\usage{
shared.data.import(
  file.name,
  excluded.columns = c("STUDYID", "DOMAIN"),
  required.columns = character(),
  dtplyr.step = FALSE,
  immutable = FALSE
)
}
\arguments{
\item{file.name}{Path of the data file (CDISC format)}

\item{excluded.columns}{Columns to be removed}

\item{dtplyr.step}{If FALSE, the processed tibble will be called \code{as_tibble} or \code{as.data.table}, otherwise not.}

\item{immutable}{If False the dataframe is immutable, i.e. it changes. This is what we
want.}
}
\value{
The contents of \code{file.name} as a tibble or \code{dtplyr_step}
}
\description{
Helper function that undertakes the first pre-processing steps common
to all RAW files provided by ISARIC.
}
\details{
First, \code{required.columns} are included (even if empty). The file is cast
as a mutable dtplyr \code{lazy_tb}. Sensitive \code{excluded.columns} are removed.
All column names are transformed to lowercase.
}
\keyword{internal}
