% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{status.by.time.after.admission.prep}
\alias{status.by.time.after.admission.prep}
\title{Aggregate data for timeline plot}
\usage{
status.by.time.after.admission.prep(input.tbl)
}
\arguments{
\item{input.tbl}{Input tibble (output of \code{data.preprocessing})}
}
\value{
A \code{tibble} containing the input data for the timeline plot
}
\description{
Aggregate data for timeline plot
}
