% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{symptom.prevalence.prep}
\alias{symptom.prevalence.prep}
\title{Aggregate data for symptom prevalence plot}
\usage{
symptom.prevalence.prep(input.tbl)
}
\arguments{
\item{input.tbl}{Input tibble (output of \code{data.preprocessing})}
}
\value{
A \code{tibble} containing the input data for the symptom prevalence plot
}
\description{
Aggregate data for symptom prevalence plot
}
