% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{symptom.upset.prep}
\alias{symptom.upset.prep}
\title{Aggregate data for symptoms upset plot}
\usage{
symptom.upset.prep(input.tbl, max.symptoms = 5)
}
\arguments{
\item{input.tbl}{Input tibble (output of \code{data.preprocessing})}

\item{max.symptoms}{The plot will display only the n most common symptoms, this parameter is n}
}
\value{
A \code{tibble} containing the input data for the symptoms upset plot
}
\description{
Aggregate data for symptoms upset plot
}
