% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{symptoms.prep}
\alias{symptoms.prep}
\title{Prepare Table3. symptoms}
\usage{
symptoms.prep(input.tbl)
}
\arguments{
\item{input.tbl}{Input tibble (output of \code{data.preprocessing})}
}
\value{
A \code{tibble} containing the input data for the Patient characteristics table
}
\description{
Prepare Table3. symptoms
}
