% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SimulatedDataFunctions.R
\name{synthesise}
\alias{synthesise}
\title{Generate a basic synthetic dataset based on a dataset supplied which maintains key relationships.}
\usage{
synthesise(data, var_identifier = "usubjid")
}
\arguments{
\item{data}{Dataframe to be used as a basis for a synthetic dataset.}

\item{var_identifier}{Vector of names of variables that are identifying variables (default = "usubjid")}
}
\value{
data.frame with values synthesised based on the original dataset
}
\description{
Used to generate an easily sharable data dictionary for an R dataframe. This supports the following classes: numeric, integer, logical, Date, character, String, factor, ordered.
}
