% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{treatment.use.proportion.prep}
\alias{treatment.use.proportion.prep}
\title{Aggregate data for treatment use proportion plot}
\usage{
treatment.use.proportion.prep(input.tbl)
}
\arguments{
\item{input.tbl}{Input tibble (output of \code{data.preprocessing})}
}
\value{
A \code{tibble} containing the input data for the treatment use proportion plot
}
\description{
Aggregate data for treatment use proportion plot
}
