% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{treatments.prep}
\alias{treatments.prep}
\title{Prepare Table5. Prevalence of treatments}
\usage{
treatments.prep(input.tbl)
}
\arguments{
\item{input.tbl}{Input tibble (output of \code{data.preprocessing})}
}
\value{
A \code{tibble} containing the input data for the Patient characteristics table
}
\description{
Prepare Table5. Prevalence of treatments
}
