% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFunctions.R, R/PreprocessingFunction.R
\docType{data}
\name{date_pull}
\alias{date_pull}
\title{Shared pre-processing of input CSV files}
\format{
An object of class \code{Date} of length 1.

An object of class \code{Date} of length 1.

An object of class \code{Date} of length 1.
}
\usage{
date_pull

date_pull

date_pull
}
\arguments{
\item{excluded.columns}{Columns to be removed}

\item{file.name}{Path of the microbio data file (CDISC format)}

\item{dtplyr.step}{Return the output as \code{dtplyr_step} to avoid unnecessary future calls to \code{as_tibble} or \code{as.data.table}}

\item{input.tbl}{Input tibble (output of \code{process.all.data})}
}
\value{
The contents of \code{file.name} as a tibble or \code{dtplyr_step}

Formatted demographic data as a tibble or \code{dtplyr_step}

A \code{tibble} intended for input into other aggregation functions (e.g. \code{age.pyramid.prep})
}
\description{
Helper function that undertakes the first pre-processing steps common
to all RAW files provided by ISARIC.


}
\details{



}
\keyword{datasets}
\keyword{internal}
