% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFunctions.R
\name{import.demographic.data}
\alias{import.demographic.data}
\title{Import demographic data}
\usage{
import.demographic.data(file.name, dtplyr.step = TRUE, date_pull = Sys.Date())
}
\arguments{
\item{file.name}{Path of the demographics data file (CDISC format)}

\item{dtplyr.step}{If FALSE, the processed tibble will be called \code{as_tibble} or \code{as.data.table}, otherwise not.}

\item{date_pull}{Date the CDISC data were provided, default is current date. Data beyond this date is discarded.}
}
\value{
Formatted demographic data as a tibble or \code{dtplyr_step}
}
\description{
Processing for the demographic data
}
\details{

}
