% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImportFunctions.R
\name{shared.data.import}
\alias{shared.data.import}
\title{Shared pre-processing of input CSV files}
\usage{
shared.data.import(
  file.name,
  excluded.columns = c("STUDYID", "DOMAIN"),
  required.columns = character(),
  dtplyr.step = FALSE,
  immutable = FALSE
)
}
\arguments{
\item{file.name}{Path of the data file (CDISC format)}

\item{excluded.columns}{Columns to be removed}

\item{dtplyr.step}{Return the output as \code{dtplyr_step} to avoid unnecessary future calls to \code{as_tibble} or \code{as.data.table}}
}
\value{
The contents of \code{file.name} as a tibble or \code{dtplyr_step}
}
\description{
Shared pre-processing of input CSV files
}
\keyword{internal}
