% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AggregationFunctions.R
\name{summary.input.prep}
\alias{summary.input.prep}
\title{Data for the report summary}
\usage{
\method{summary}{input.prep}(input.tbl)
}
\arguments{
\item{input.tbl}{Input tibble (output of \code{data.preprocessing})}
}
\value{
A \code{tibble} containing the input data for the symptoms upset plot
}
\description{
Data for the report summary
}
